refactor train.py and train_lora.py
